perf(bundle): lazy-load proposal-panels subsystem out of entry chunk
Extracted ~1500 lines / ~80 kB raw of cold-path proposal rendering
(goal/role/tool/staff/project preview + proposal panels, renderGoalForm,
workflow/component/qa loaders, triggers editor wrapper) into a new
src/app/proposal-panels.ts module. The entry chunk imports a thin
proposal-panels-lazy.ts wrapper that dynamic-imports the heavy module
on first proposal-panel view; until then a bouncing-bobbit placeholder
shows for ~50ms. Fire-and-forget actions (setSelectedWorkflowId,
showProposalToast, dismissTypedProposal, etc.) queue and drain when
the chunk lands.

Drops entry chunk from 617.91 kB raw / 161.86 kB gz to 570.44 kB raw /
150.24 kB gz, comfortably under the 600 kB Vite warning threshold.
Spawns a new proposal-panels-*.js chunk at 50.29 kB raw / 13.46 kB gz.
